Default to using control sockets for SSH persisting 15 minutes
This then requires only one auth for SSH each backup.  This is
convenient if you auth with a password or some key that isn't in root's
ssh-agent.

If you backup more often than the default ControlPersist (say, in a loop
every 10 minutes), you can auth once you start you internet connection
and then it will reuse it for every subsequent backup.

This is supported since 2004 on the client side for OpenSSH.